Fixed compilation in VS-2015 * compile failed for 32 bit platforms * missing include algorithm header * std::numeric_limits<T>::max conflict with some MS header defines
