+++ lib/tinyutf8.cpp
@@ -26,6 +26,7 @@
 #include "tinyutf8.h"
 #include <ostream>
 #include <istream>
+#include <algorithm>
 
 utf8_string::utf8_string( utf8_string::size_type count , utf8_string::value_type cp ) :
 	t_sso( 0 )
@@ -425,6 +426,16 @@ utf8_string::width_type utf8_string::get_num_bytes_of_utf8_char_before( const ch
 	}
 }
 
+
+utf8_string::width_type utf8_string::get_lut_width(size_type buffer_size) {
+	return buffer_size <= std::numeric_limits<std::uint8_t>::max()
+			? sizeof(std::uint8_t)
+			: buffer_size <= std::numeric_limits<std::uint16_t>::max()
+				? sizeof(std::uint16_t)
+				: buffer_size <= std::numeric_limits<std::uint32_t>::max()
+					? sizeof(std::uint32_t)
+					: sizeof(std::uint64_t);
+}
 
 
 #if !defined(_TINY_UTF8_H_HAS_CLZ_) || _TINY_UTF8_H_HAS_CLZ_ == false
